third_party/WebKit/LayoutTests/http/tests/devtools/elements/dom-agent-query-selector-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests DOMAgent.querySelector and DOMAgent.querySelectorAll.


Running: testDocumentQuerySelector
node id: id1

Running: testDocumentQuerySelectorAll
node id: id1
node id: id2
node id: id3
node id: id4
node id: id5
node id: id6

Running: testQuerySelector
node id: id3

Running: testQuerySelectorAll
node id: id3
node id: id4
node id: id5
node id: id6
